requirements.txt
celery==5.2.7
django==4.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-extensions==3.2.0
django-filter==23.1
django-environ==0.10.0
django-storages==1.13.2
django-watchman==1.3.0
dj-database-url==1.0.0
drf-jwt==1.19.2
drf-spectacular==0.26.1
ipython==8.4.0
psycopg2-binary==2.9.5
redis==4.5.4
uvicorn==0.21.1
